move sample user document to own file
